backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.23",
    "fastapi>=0.129.0",
    "huggingface-hub>=0.34.0",
    "numpy>=2.3.0",
    "pymupdf>=1.26.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "torch>=2.8.0",
    "transformers>=4.57.0",
    "uvicorn[standard]>=0.41.0",
    "torchaudio>=2.10.0",
    "websockets>=15.0.1",
    "docx2txt>=0.9",
]
